feat: add an option to download the source configuration (#11)
* feat: add controlPlaneUrl in configuration

* chore: fix the buildSrc/RudderStackBuildConfig import

This involves creating the `src` folder and also updating few fields.

* chore: add options to store version code and name in both Android and Core modules

These values is taken from RudderStackBuildConfig class.

* feat: add option to get the LibraryVersion in Storage

This needs to be implemented by every Storage class. This info is needed at two places: first at the context of each payload and second at for fetching SourceConfig

* feat: add SourceConfig data class

* test: add unit test cases for SourceConfig

* feat: add SourceConfigManager.kt

It'll download the sourceConfig. I've also added a TODO.

* feat: integrate SourceConfig component in core module

* refactor: make SourceConfigManager class unit testable

Also added a test class which will be completed once we merge all the PR's.

* fix: make destination object as non-nullable

Set empty list as its default value

* fix: make metric object as non-nullable

Provide the default value for this object. Now by default metrics will be disabled.

* test: remove destination config from test config files

* test: create util method to read file as string

* test: create provider for sourceConfig

* refactor(test): utilise provider in SourceConfigTest

* test: remove incomplete test code

* chore: reverse the validateBaseUrl logic

Now, if '/' is added as suffix in the URL remove it.

* fix: add `/` as a prefix for sourceConfig endpoint