feat: implement track api (#18)
* chore: remove equal and hashcode methods from Message

As of now, we aren't sure where exactly we need them. So, I cleaned up the code. I've kept copy as it'll be required while sending events to the integrations

* refactor: track event parameter

* chore: remove newId

* chore: add encodeDefaults support in JsonUtils

* fix: make the options field non-serializable

We need to make their properties serializable i.e., integrations and externalIds. There is no field called options in the event payload.

* feat: add integrations field in Message

It is currently set to default value.

* refactor: rename from `name` to `event` field

Event is defined in our spec.

* chore: allow forbidden comments

This is necessary as of now to keep track of pending items. We can revert this once we release the beta SDK.

* feat: add anonymousId support in Message

* feat: add integrations and anonymousId in copy method

* feat: add encodeToString extension function

This will be helpful to convert Message object into String. This will also filter the message object.

* doc: remove property definition

This changes is done to address the compiler warning.

* feat: introduce interface for Platform

* feat: implement Platform type in both Analytic modules

Android will provide "mobile" and kotlin will provide "server" as their platform types.

* feat: add channel type support in Message

* feat: make `option` an optional field in track event

* refactor: RudderOption to a new class

This is to prevent cluttering the code in the Message class. And also to add other methods related to options class.

* fix: handling of integrations option

* fix: revert back to String -> Boolean type for integrations option

* feat: add util functions to improve readability

* feat: add customContext support and improve readability

I've added the infix util functions to perform merging operation and I'm aiming that this functions will assist in readability.

* test: add all expected JSON files for track event

* test: add test cases for track event

* refactor(test): change the test class name to TrackEventTest

* chore: restrict visibility of utility methods to internal

* chore: address comment