chore: improve kotlin sdk (#99)
* chore(android): remove stale delete from SharedPrefs

Currently, this method is not utilised.

* chore(android): remove stale putFloat from SharedPrefs

Currently there is no API to save Float values.

* chore(android): change modifier of DIRECTORY_NAME to private

This was previously set to internal but not used anywhere else other than the current class.

* chore(core): replace property with param in BasicStorage

* chore(android): correct the reference of Configuration in comments

* chore(android): correct the reference of Analytics in comments

* chore: make CheckBuildVersionUseCase.isAndroidVersionLollipopAndAbove internal

* chore: make AndroidStorage provideAndroidStorage() internal

* chore: make BasicStorage provideBasicStorage() internal

* chore: mark multiple SessionTrackingPlugin method as internal

Note: This is just to improve the readability. As the class is internal, we are not really changing anything.

* chore: mark NavControllerActivityObserver.find() as internal

Note: This is just to improve the readability. As the class is internal, we are not really changing anything.

* chore: mark multiple ActivityLifecycleManagementPlugin as internal

Note: This is just to improve the readability. As the class is internal, we are not really changing anything.

* chore: mark multiple ProcessLifecycleManagementPlugin method as internal

Note: This is just to improve the readability. As the class is internal, we are not really changing anything.

* chore(core): mark AnalyticsConfiguration.provideAnalyticsConfiguration with InternalRudderApi annotation

* chore(core): restrict public visibility of StringUtils.kt

By either marking method as internal or adding InternalRudderApi annotation

* chore(core): restrict public visibility of JSONUtils.kt

By either marking method as internal or adding InternalRudderApi annotation

* chore(core): restrict public visibility of JsonSentAtUpdater.kt

By either marking method as internal or adding InternalRudderApi annotation

* chore(core): restrict public visibility of DateTimeUtils.kt

By either marking method as internal or adding InternalRudderApi annotation

* refactor: AndroidAdvertisingPlugin to decouple it from sdk util methods

Currently, SDK cannot change it's util method, if user get dependent on this.

* chore(core): restrict public visibility of Base64.kt methods

By either marking method as internal or adding InternalRudderApi annotation

* chore(core): restrict public visibility of AnalyticsUtils.kt methods

By either marking method as internal or adding InternalRudderApi annotation

* chore(core): restrict public visibility of PropertiesFile.kt methods

By either marking method as internal or adding InternalRudderApi annotation

* chore(core): restrict public visibility of EventBatchFileManager.kt

By either marking method as internal or adding InternalRudderApi annotation

* chore: remove UnsupportedTypeException.kt

As this class is not used anywhere.

* chore(core): restrict public visibility of FlowState.kt methods

By either marking method as internal or adding InternalRudderApi annotation

* chore(core): restrict public visibility of EventQueue.kt methods

By either marking method as internal or adding InternalRudderApi annotation

* chore(core): restrict public visibility of PlatformType

By either marking method as internal or adding InternalRudderApi annotation

* chore(core): restrict public visibility of HttpClientImpl.kt and test class

By either marking method as internal or adding InternalRudderApi annotation

* chore: remove unnecessary @InternalRudderApi annotation from DateTimeUtils

* refactor: replace use of putAll with mergeWithHigherPriorityTo

* chore: change the visibility modifier of generateUUID

This is to ensure that this method can be utilised in future even in android module.